+++ shared/observability/tests/observability/metrics/test_stats.py
@@ -27,7 +27,6 @@ from unittest.mock import Mock
 import pytest
 import statsd
 
-import airflow._shared.observability.stats
 from airflow._shared.configuration.exceptions import AirflowConfigException
 from airflow._shared.observability.exceptions import InvalidStatsNameException
 from airflow._shared.observability.metrics.datadog_logger import SafeDogStatsdLogger
@@ -36,6 +35,7 @@ from airflow._shared.observability.metrics.validators import (
     PatternAllowListValidator,
     PatternBlockListValidator,
 )
+from airflow.observability import stats
 
 from tests_common.test_utils.config import conf_vars
 from tests_common.test_utils.markers import skip_if_force_lowest_dependencies_marker
@@ -102,11 +102,11 @@ class TestStats:
     def test_enabled_by_config(self):
         """Test that enabling this sets the right instance properties"""
         with conf_vars({("metrics", "statsd_on"): "True"}):
-            importlib.reload(airflow._shared.observability.stats)
-            assert isinstance(airflow._shared.observability.stats.Stats.statsd, statsd.StatsClient)
-            assert not hasattr(airflow._shared.observability.stats.Stats, "dogstatsd")
+            importlib.reload(stats)
+            assert isinstance(stats.Stats.statsd, statsd.StatsClient)
+            assert not hasattr(stats.Stats, "dogstatsd")
         # Avoid side-effects
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
 
     def test_load_custom_statsd_client(self):
         with conf_vars(
@@ -115,10 +115,10 @@ class TestStats:
                 ("metrics", "statsd_custom_client_path"): f"{__name__}.CustomStatsd",
             }
         ):
-            importlib.reload(airflow._shared.observability.stats)
-            assert isinstance(airflow._shared.observability.stats.Stats.statsd, CustomStatsd)
+            importlib.reload(stats)
+            assert isinstance(stats.Stats.statsd, CustomStatsd)
         # Avoid side-effects
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
 
     def test_load_invalid_custom_stats_client(self):
         with conf_vars(
@@ -127,14 +127,14 @@ class TestStats:
                 ("metrics", "statsd_custom_client_path"): f"{__name__}.InvalidCustomStatsd",
             }
         ):
-            importlib.reload(airflow._shared.observability.stats)
+            importlib.reload(stats)
             error_message = re.escape(
                 "Your custom StatsD client must extend the statsd."
                 "StatsClient in order to ensure backwards compatibility."
             )
             with pytest.raises(AirflowConfigException, match=error_message):
-                airflow._shared.observability.stats.Stats.incr("empty_key")
-        importlib.reload(airflow._shared.observability.stats)
+                stats.Stats.incr("empty_key")
+        importlib.reload(stats)
 
     def test_load_allow_list_validator(self):
         with conf_vars(
@@ -143,16 +143,14 @@ class TestStats:
                 ("metrics", "metrics_allow_list"): "name1,name2",
             }
         ):
-            importlib.reload(airflow._shared.observability.stats)
-            assert (
-                type(airflow._shared.observability.stats.Stats.metrics_validator) is PatternAllowListValidator
-            )
-            assert airflow._shared.observability.stats.Stats.metrics_validator.validate_list == (
+            importlib.reload(stats)
+            assert type(stats.Stats.metrics_validator) is PatternAllowListValidator
+            assert stats.Stats.metrics_validator.validate_list == (
                 "name1",
                 "name2",
             )
         # Avoid side-effects
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
 
     def test_load_block_list_validator(self):
         with conf_vars(
@@ -161,16 +159,14 @@ class TestStats:
                 ("metrics", "metrics_block_list"): "name1,name2",
             }
         ):
-            importlib.reload(airflow._shared.observability.stats)
-            assert (
-                type(airflow._shared.observability.stats.Stats.metrics_validator) is PatternBlockListValidator
-            )
-            assert airflow._shared.observability.stats.Stats.metrics_validator.validate_list == (
+            importlib.reload(stats)
+            assert type(stats.Stats.metrics_validator) is PatternBlockListValidator
+            assert stats.Stats.metrics_validator.validate_list == (
                 "name1",
                 "name2",
             )
         # Avoid side-effects
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
 
     def test_load_allow_and_block_list_validator_loads_only_allow_list_validator(self):
         with conf_vars(
@@ -180,16 +176,14 @@ class TestStats:
                 ("metrics", "metrics_block_list"): "name1,name2",
             }
         ):
-            importlib.reload(airflow._shared.observability.stats)
-            assert (
-                type(airflow._shared.observability.stats.Stats.metrics_validator) is PatternAllowListValidator
-            )
-            assert airflow._shared.observability.stats.Stats.metrics_validator.validate_list == (
+            importlib.reload(stats)
+            assert type(stats.Stats.metrics_validator) is PatternAllowListValidator
+            assert stats.Stats.metrics_validator.validate_list == (
                 "name1",
                 "name2",
             )
         # Avoid side-effects
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
 
 
 class TestDogStats:
@@ -275,11 +269,11 @@ class TestDogStats:
         from datadog import DogStatsd
 
         with conf_vars({("metrics", "statsd_datadog_enabled"): "True"}):
-            importlib.reload(airflow._shared.observability.stats)
-            assert isinstance(airflow._shared.observability.stats.Stats.dogstatsd, DogStatsd)
-            assert not hasattr(airflow._shared.observability.stats.Stats, "statsd")
+            importlib.reload(stats)
+            assert isinstance(stats.Stats.dogstatsd, DogStatsd)
+            assert not hasattr(stats.Stats, "statsd")
         # Avoid side-effects
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
 
     def test_does_not_send_stats_using_statsd_when_statsd_and_dogstatsd_both_on(self):
         from datadog import DogStatsd
@@ -290,10 +284,10 @@ class TestDogStats:
                 ("metrics", "statsd_datadog_enabled"): "True",
             }
         ):
-            importlib.reload(airflow._shared.observability.stats)
-            assert isinstance(airflow._shared.observability.stats.Stats.dogstatsd, DogStatsd)
-            assert not hasattr(airflow._shared.observability.stats.Stats, "statsd")
-        importlib.reload(airflow._shared.observability.stats)
+            importlib.reload(stats)
+            assert isinstance(stats.Stats.dogstatsd, DogStatsd)
+            assert not hasattr(stats.Stats, "statsd")
+        importlib.reload(stats)
 
 
 class TestStatsAllowAndBlockLists:
@@ -352,7 +346,7 @@ class TestStatsAllowAndBlockLists:
 class TestPatternValidatorConfigOption:
     def teardown_method(self):
         # Avoid side-effects
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
 
     stats_on = {("metrics", "statsd_on"): "True"}
     allow_list = {("metrics", "metrics_allow_list"): "foo,bar"}
@@ -385,21 +379,18 @@ class TestPatternValidatorConfigOption:
     )
     def test_pattern_picker(self, config, expected):
         with conf_vars(config):
-            importlib.reload(airflow._shared.observability.stats)
+            importlib.reload(stats)
 
-            assert isinstance(airflow._shared.observability.stats.Stats.statsd, statsd.StatsClient)
-            assert type(airflow._shared.observability.stats.Stats.instance.metrics_validator) is expected
+            assert isinstance(stats.Stats.statsd, statsd.StatsClient)
+            assert type(stats.Stats.instance.metrics_validator) is expected
 
     @conf_vars({**stats_on, **block_list, ("metrics", "metrics_allow_list"): "baz,qux"})
     def test_setting_allow_and_block_logs_warning(self, caplog):
         with caplog.at_level(logging.WARNING):
-            importlib.reload(airflow._shared.observability.stats)
+            importlib.reload(stats)
 
-            assert isinstance(airflow._shared.observability.stats.Stats.statsd, statsd.StatsClient)
-            assert (
-                type(airflow._shared.observability.stats.Stats.instance.metrics_validator)
-                is PatternAllowListValidator
-            )
+            assert isinstance(stats.Stats.statsd, statsd.StatsClient)
+            assert type(stats.Stats.instance.metrics_validator) is PatternAllowListValidator
             assert "Ignoring metrics_block_list" in caplog.text
 
 
@@ -521,8 +512,8 @@ class TestCustomStatsName:
     )
     @mock.patch("statsd.StatsClient")
     def test_does_not_send_stats_using_statsd_when_the_name_is_not_valid(self, mock_statsd):
-        importlib.reload(airflow._shared.observability.stats)
-        airflow._shared.observability.stats.Stats.incr("empty_key")
+        importlib.reload(stats)
+        stats.Stats.incr("empty_key")
         mock_statsd.return_value.assert_not_called()
 
     @skip_if_force_lowest_dependencies_marker
@@ -534,8 +525,8 @@ class TestCustomStatsName:
     )
     @mock.patch("datadog.DogStatsd")
     def test_does_not_send_stats_using_dogstatsd_when_the_name_is_not_valid(self, mock_dogstatsd):
-        importlib.reload(airflow._shared.observability.stats)
-        airflow._shared.observability.stats.Stats.incr("empty_key")
+        importlib.reload(stats)
+        stats.Stats.incr("empty_key")
         mock_dogstatsd.return_value.assert_not_called()
 
     @conf_vars(
@@ -546,8 +537,8 @@ class TestCustomStatsName:
     )
     @mock.patch("statsd.StatsClient")
     def test_does_send_stats_using_statsd_when_the_name_is_valid(self, mock_statsd):
-        importlib.reload(airflow._shared.observability.stats)
-        airflow._shared.observability.stats.Stats.incr("empty_key")
+        importlib.reload(stats)
+        stats.Stats.incr("empty_key")
         mock_statsd.return_value.incr.assert_called_once_with("empty_key", 1, 1)
 
     @skip_if_force_lowest_dependencies_marker
@@ -559,12 +550,12 @@ class TestCustomStatsName:
     )
     @mock.patch("datadog.DogStatsd")
     def test_does_send_stats_using_dogstatsd_when_the_name_is_valid(self, mock_dogstatsd):
-        importlib.reload(airflow._shared.observability.stats)
-        airflow._shared.observability.stats.Stats.incr("empty_key")
+        importlib.reload(stats)
+        stats.Stats.incr("empty_key")
         mock_dogstatsd.return_value.increment.assert_called_once_with(
             metric="empty_key", sample_rate=1, tags=[], value=1
         )
 
     def teardown_method(self) -> None:
         # To avoid side-effect
-        importlib.reload(airflow._shared.observability.stats)
+        importlib.reload(stats)
